install-headers target should depend on BUILT_SOURCES
Quoting from https://www.gnu.org/software/automake/manual/html_node/Sources.html

    BUILT_SOURCES is actually a bit of a misnomer, as any file which must
    be created early in the build process can be listed in this
    variable. Moreover, all built sources do not necessarily have to be
    listed in BUILT_SOURCES. For instance, a generated .c file doesn’t
    need to appear in BUILT_SOURCES (unless it is included by another
    source), because it’s a known dependency of the associated object.

The documentation for BUILT_SOURCES repeatedly states that the it
specifies the sources that need to be built early in the build process
and that `make all`, `make check` and `make install` all depend on
`BUILT_SOURCES`.  Given the strong parallels between those commands
and `make install-headers` we argue that introducing this dependency
is a reasonable step in enforcing the logical consistency of the build
system.