Remove inconsistent libtool libraries and automake-distributed files
The m4 files removed in this pull request are spurious in the
repository.  When the developer chooses to to build the toolchain
pinned by the repository, these exact files get pulled down as a part
of the libtool package and they are used appropriately in the
bootstrap process. In that case, the removed files are spurious.  When
the developer chooses to use the system's version of libtool and
automake, these file confuse the bootstrap and cause the libtool to
not work correctly on the initial bootstrap.  In that case, these
files are actively harmful.